+++ packages/features/data-table/hooks/useFilterableColumns.ts
@@ -4,12 +4,14 @@ import { type Table } from "@tanstack/react-table";
 // eslint-disable-next-line no-restricted-imports
 import { useMemo } from "react";
 
-import type { FilterableColumn } from "../lib/types";
+import type { FilterableColumn, FacetedValue } from "../lib/types";
 import { ColumnFilterType } from "../lib/types";
+import { convertMapToFacetedValues } from "../lib/utils";
 
 export function useFilterableColumns<TData>(table: Table<TData>) {
   const columns = useMemo(
     () => table.getAllColumns().filter((column) => column.getCanFilter()),
+
     [table.getAllColumns()]
   );
 
@@ -25,20 +27,18 @@ export function useFilterableColumns<TData>(table: Table<TData>) {
             type,
           };
           if (type === ColumnFilterType.MULTI_SELECT || type === ColumnFilterType.SINGLE_SELECT) {
-            const values = column.getFacetedUniqueValues();
-            const options = Array.from(values.keys()).map((option) => {
-              if (typeof option === "string") {
-                return {
-                  label: option,
-                  value: option,
-                };
-              } else {
-                return {
-                  label: option.label as string,
-                  value: option.value as string | number,
-                };
-              }
-            });
+            // `column.getFacetedUniqueValues` gets out of sync
+            // when we pass a new `getFacetedUniqueValues` to
+            // `useReactTable({ ... })`.
+            //
+            // So we use `table.options.getFacetedUniqueValues` instead.
+            let values: Map<FacetedValue, number> | (() => Map<FacetedValue, number>) | undefined =
+              table.options?.getFacetedUniqueValues?.(table, column.id);
+            if (typeof values === "function") {
+              values = values();
+            }
+
+            const options = convertMapToFacetedValues(values);
             return {
               ...base,
               options,
@@ -50,7 +50,10 @@ export function useFilterableColumns<TData>(table: Table<TData>) {
           }
         })
         .filter((column): column is FilterableColumn => Boolean(column)),
-    [columns]
+
+    // re-calculate this when the `getFacetedUniqueValues` changes
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+    [columns, table.options.getFacetedUniqueValues]
   );
 
   return filterableColumns;
