+++ packages/features/data-table/lib/types.ts
@@ -148,3 +148,8 @@ export const ZSorting = z.object({
 export type SortingState = Array<z.infer<typeof ZSorting>>;
 
 export const ZColumnVisibility = z.record(z.string(), z.boolean());
+
+export type FacetedValue = {
+  label: string;
+  value: string | number;
+};
