+++ packages/features/filters/components/TeamsFilter.tsx
@@ -154,38 +154,43 @@ export const FilterCheckboxFieldsContainer = ({
 
 type Props = InputHTMLAttributes<HTMLInputElement> & {
   label: string;
+  testId?: string;
   icon?: ReactNode;
 };
 
-export const FilterCheckboxField = forwardRef<HTMLInputElement, Props>(({ label, icon, ...rest }, ref) => {
-  return (
-    <div className="hover:bg-muted flex items-center py-2 pl-3 pr-2.5 transition hover:cursor-pointer">
-      <label className="flex w-full max-w-full items-center justify-between hover:cursor-pointer">
-        <div className="flex items-center truncate">
-          {icon && (
-            <div className="text-default flex h-4 w-4 items-center justify-center ltr:mr-2 rtl:ml-2">
-              {icon}
-            </div>
-          )}
-          <Tooltip content={label}>
-            <label
-              htmlFor={rest.id}
-              className="text-default me-1 cursor-pointer truncate text-sm font-medium">
-              {label}
-            </label>
-          </Tooltip>
-        </div>
-        <div className="flex h-5 items-center">
-          <input
-            {...rest}
-            ref={ref}
-            type="checkbox"
-            className="text-emphasis dark:text-muted focus:ring-emphasis border-default bg-default h-4 w-4 rounded transition hover:cursor-pointer"
-          />
-        </div>
-      </label>
-    </div>
-  );
-});
+export const FilterCheckboxField = forwardRef<HTMLInputElement, Props>(
+  ({ label, icon, testId, ...rest }, ref) => {
+    return (
+      <div
+        data-testid={testId}
+        className="hover:bg-muted flex items-center py-2 pl-3 pr-2.5 transition hover:cursor-pointer">
+        <label className="flex w-full max-w-full items-center justify-between hover:cursor-pointer">
+          <div className="flex items-center truncate">
+            {icon && (
+              <div className="text-default flex h-4 w-4 items-center justify-center ltr:mr-2 rtl:ml-2">
+                {icon}
+              </div>
+            )}
+            <Tooltip content={label}>
+              <label
+                htmlFor={rest.id}
+                className="text-default me-1 cursor-pointer truncate text-sm font-medium">
+                {label}
+              </label>
+            </Tooltip>
+          </div>
+          <div className="flex h-5 items-center">
+            <input
+              {...rest}
+              ref={ref}
+              type="checkbox"
+              className="text-emphasis dark:text-muted focus:ring-emphasis border-default bg-default h-4 w-4 rounded transition hover:cursor-pointer"
+            />
+          </div>
+        </label>
+      </div>
+    );
+  }
+);
 
 FilterCheckboxField.displayName = "FilterCheckboxField";
