+++ packages/features/insights/components/LeastBookedTeamMembersTable.tsx
@@ -3,23 +3,21 @@ import { Title } from "@tremor/react";
 import { useLocale } from "@calcom/lib/hooks/useLocale";
 import { trpc } from "@calcom/trpc";
 
-import { useFilterContext } from "../context/provider";
+import { useInsightsParameters } from "../hooks/useInsightsParameters";
 import { CardInsights } from "./Card";
 import { LoadingInsight } from "./LoadingInsights";
 import { TotalBookingUsersTable } from "./TotalBookingUsersTable";
 
 export const LeastBookedTeamMembersTable = () => {
   const { t } = useLocale();
-  const { filter } = useFilterContext();
-  const { dateRange, selectedEventTypeId, selectedTeamId: teamId, isAll, initialConfig } = filter;
-  const [startDate, endDate] = dateRange;
+  const { isAll, teamId, startDate, endDate, eventTypeId } = useInsightsParameters();
 
   const { data, isSuccess, isPending } = trpc.viewer.insights.membersWithLeastBookings.useQuery(
     {
-      startDate: startDate.toISOString(),
-      endDate: endDate.toISOString(),
+      startDate,
+      endDate,
       teamId,
-      eventTypeId: selectedEventTypeId ?? undefined,
+      eventTypeId,
       isAll,
     },
     {
@@ -27,13 +25,12 @@ export const LeastBookedTeamMembersTable = () => {
       trpc: {
         context: { skipBatch: true },
       },
-      enabled: !!(initialConfig?.teamId || initialConfig?.userId || initialConfig?.isAll),
     }
   );
 
   if (isPending) return <LoadingInsight />;
 
-  if (!isSuccess || !startDate || !endDate || !teamId) return null;
+  if (!isSuccess || !data) return null;
 
   return (
     <CardInsights>
