+++ packages/features/insights/components/LowestRatedMembersTable.tsx
@@ -3,24 +3,21 @@ import { Title } from "@tremor/react";
 import { useLocale } from "@calcom/lib/hooks/useLocale";
 import { trpc } from "@calcom/trpc";
 
-import { useFilterContext } from "../context/provider";
+import { useInsightsParameters } from "../hooks/useInsightsParameters";
 import { CardInsights } from "./Card";
 import { LoadingInsight } from "./LoadingInsights";
 import { TotalUserFeedbackTable } from "./TotalUserFeedbackTable";
 
 export const LowestRatedMembersTable = () => {
   const { t } = useLocale();
-  const { filter } = useFilterContext();
-  const { dateRange, selectedEventTypeId, isAll, initialConfig } = filter;
-  const [startDate, endDate] = dateRange;
-  const { selectedTeamId: teamId } = filter;
+  const { isAll, teamId, startDate, endDate, eventTypeId } = useInsightsParameters();
 
   const { data, isSuccess, isPending } = trpc.viewer.insights.membersWithLowestRatings.useQuery(
     {
-      startDate: startDate.toISOString(),
-      endDate: endDate.toISOString(),
+      startDate,
+      endDate,
       teamId,
-      eventTypeId: selectedEventTypeId ?? undefined,
+      eventTypeId,
       isAll,
     },
     {
@@ -28,13 +25,12 @@ export const LowestRatedMembersTable = () => {
       trpc: {
         context: { skipBatch: true },
       },
-      enabled: !!(initialConfig?.teamId || initialConfig?.userId || initialConfig?.isAll),
     }
   );
 
   if (isPending) return <LoadingInsight />;
 
-  if (!isSuccess || !startDate || !endDate || !teamId) return null;
+  if (!isSuccess || !data) return null;
 
   return data && data.length > 0 ? (
     <CardInsights className="shadow-none">
