+++ packages/features/insights/filters/Download/Download.tsx
@@ -1,22 +1,23 @@
-import { useFilterContext } from "@calcom/features/insights/context/provider";
 import { downloadAsCsv } from "@calcom/lib/csvUtils";
 import { useLocale } from "@calcom/lib/hooks/useLocale";
 import type { RouterOutputs } from "@calcom/trpc";
 import { trpc } from "@calcom/trpc";
 import { Button, Dropdown, DropdownItem, DropdownMenuContent, DropdownMenuTrigger } from "@calcom/ui";
 
+import { useInsightsParameters } from "../../hooks/useInsightsParameters";
+
 const Download = () => {
-  const { filter } = useFilterContext();
   const { t } = useLocale();
+  const { startDate, endDate, teamId, userId, eventTypeId, memberUserId } = useInsightsParameters();
 
   const { data, isPending } = trpc.viewer.insights.rawData.useQuery(
     {
-      startDate: filter.dateRange[0].toISOString(),
-      endDate: filter.dateRange[1].toISOString(),
-      teamId: filter.selectedTeamId,
-      userId: filter.selectedUserId,
-      eventTypeId: filter.selectedEventTypeId,
-      memberUserId: filter.selectedMemberUserId,
+      startDate,
+      endDate,
+      teamId,
+      userId,
+      eventTypeId,
+      memberUserId,
     },
     {
       refetchOnWindowFocus: false,
