+++ packages/features/insights/hooks/useInsightsOrgTeams.ts
@@ -8,10 +8,11 @@ export function useInsightsOrgTeams() {
   if (!context) {
     throw new Error("useInsightsOrgTeams must be used within a InsightsOrgTeamsProvider");
   }
-  const { orgTeamsType, selectedTeamId, setOrgTeamsType, setSelectedTeamId } = context;
   const session = useSession();
+  const currentOrgId = session.data?.user.org?.id;
+  const { orgTeamsType, selectedTeamId, setOrgTeamsType, setSelectedTeamId } = context;
   const isAll = orgTeamsType === "org";
-  const teamId = orgTeamsType === "team" ? selectedTeamId : undefined;
+  const teamId = orgTeamsType === "org" ? currentOrgId : orgTeamsType === "team" ? selectedTeamId : undefined;
   const userId = orgTeamsType === "yours" ? session.data?.user.id : undefined;
 
   return {
