+++ packages/features/insights/server/trpc-router.ts
@@ -987,17 +987,13 @@ export const insightsRouter = router({
   userList: authedProcedure
     .input(
       z.object({
-        teamId: z.coerce.number().nullable(),
+        teamId: z.number().optional(),
         isAll: z.boolean().nullable(),
       })
     )
     .query(async ({ ctx, input }) => {
       const user = ctx.user;
       const { teamId, isAll } = input;
-
-      if (!teamId) {
-        return [];
-      }
 
       if (isAll && user.organizationId && user.organization.isOrgAdmin) {
         const usersInTeam = await ctx.insightsDb.membership.findMany({
@@ -1014,6 +1010,10 @@ export const insightsRouter = router({
           distinct: ["userId"],
         });
         return usersInTeam.map((membership) => membership.user);
+      }
+
+      if (!teamId) {
+        return [];
       }
 
       const membership = await ctx.insightsDb.membership.findFirst({
@@ -1552,7 +1552,7 @@ export const insightsRouter = router({
     .query(async ({ ctx, input }) => {
       const { userId, teamId, isAll } = input;
       return await RoutingEventsInsights.getRoutingFormsForFilters({
-        userId,
+        userId: ctx.user.id,
         teamId,
         isAll,
         organizationId: ctx.user.organizationId ?? undefined,
@@ -1569,7 +1569,7 @@ export const insightsRouter = router({
         organizationId: ctx.user.organizationId ?? null,
         routingFormId: input.routingFormId,
         cursor: input.cursor,
-        userId: input.userId,
+        userId: ctx.user.id,
         memberUserIds: input.memberUserIds,
         limit: input.limit,
         columnFilters: input.columnFilters,
@@ -1587,7 +1587,7 @@ export const insightsRouter = router({
         organizationId: ctx.user.organizationId ?? null,
         routingFormId: input.routingFormId,
         cursor: input.cursor,
-        userId: input.userId,
+        userId: ctx.user.id,
         memberUserIds: input.memberUserIds,
         limit: input.limit,
         columnFilters: input.columnFilters,
@@ -1605,7 +1605,7 @@ export const insightsRouter = router({
         organizationId: ctx.user.organizationId ?? null,
         routingFormId: input.routingFormId,
         cursor: input.cursor,
-        userId: input.userId,
+        userId: ctx.user.id,
         memberUserIds: input.memberUserIds,
         limit: input.limit ?? BATCH_SIZE,
         columnFilters: input.columnFilters,
@@ -1624,6 +1624,7 @@ export const insightsRouter = router({
     .query(async ({ input, ctx }) => {
       const options = await RoutingEventsInsights.getRoutingFormFieldOptions({
         ...input,
+        userId: ctx.user.id,
         organizationId: ctx.user.organizationId ?? null,
       });
       return options;
@@ -1640,6 +1641,7 @@ export const insightsRouter = router({
     .query(async ({ ctx, input }) => {
       return await RoutingEventsInsights.getFailedBookingsByRoutingFormGroup({
         ...input,
+        userId: ctx.user.id,
         organizationId: ctx.user.organizationId ?? null,
       });
     }),
@@ -1654,7 +1656,7 @@ export const insightsRouter = router({
     )
     .query(async ({ ctx, input }) => {
       const headers = await RoutingEventsInsights.getRoutingFormHeaders({
-        userId: input.userId ?? null,
+        userId: ctx.user.id,
         teamId: input.teamId ?? null,
         isAll: input.isAll,
         organizationId: ctx.user.organizationId ?? null,
@@ -1683,7 +1685,7 @@ export const insightsRouter = router({
         input;
 
       return await RoutingEventsInsights.routedToPerPeriod({
-        userId: userId ?? null,
+        userId: ctx.user.id,
         teamId: teamId ?? null,
         startDate,
         endDate,
@@ -1709,7 +1711,7 @@ export const insightsRouter = router({
       const { startDate, endDate } = input;
       try {
         const csvData = await RoutingEventsInsights.routedToPerPeriodCsv({
-          userId: input.userId ?? null,
+          userId: ctx.user.id,
           teamId: input.teamId ?? null,
           startDate,
           endDate,
